chat/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==5.1.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
einops==0.8.0
exceptiongroup==1.2.2
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
inflection==0.5.1
Jinja2==3.1.3
jiter==0.8.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.1.3
openai==1.56.1
packaging==24.2
pgvector==0.3.6
pillow==10.2.0
psycopg2==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
